Assignment2/backend/requirements.txt
fastapi[standard]
uvicorn
chromadb
google-generativeai
python-multipart
pydantic
python-dotenv
Pillow
